fix: Refactor event handling logic.
