+++ src/lib/interaction/events.ts
@@ -3,10 +3,205 @@ import { get } from 'svelte/store';
 
 import { feature } from '$lib/state/feature.svelte';
 import { layerId } from '$lib/state/layerId.svelte';
-import { listeners } from '$lib/state/listeners.svelte';
+import { listeners, type LayerListeners } from '$lib/state/listeners.svelte';
+import { popup } from '$lib/state/popup.svelte';
+import { ir } from '$lib/stores/ir';
 import { layout } from '$lib/stores/layout';
 import type { CartoKitLayer } from '$lib/types';
 import { getCanonicalLayerId, getSourceId } from '$lib/utils/layer/id';
+
+/**
+ * Add a hover effect to all features in a point layer.
+ *
+ * @param map The top-level {@link maplibregl.Map} instance.
+ * @param layerId The id of the layer to instrument.
+ * @param sourceLayerId The id of the source layer to instrument. This is only
+ * necessary for {@link CartoKitLayer}s with {@link CartoKitVectorSource}s.
+ */
+export function instrumentPointHover(
+  map: maplibregl.Map,
+  layerId: string,
+  sourceLayerId?: string
+): void {
+  const currentStrokeWidth = map.getPaintProperty(
+    layerId,
+    'circle-stroke-width'
+  );
+  const currentStrokeColor = map.getPaintProperty(
+    layerId,
+    'circle-stroke-color'
+  );
+
+  map.setPaintProperty(layerId, 'circle-stroke-width', [
+    'case',
+    ['boolean', ['feature-state', 'hover'], false],
+    1,
+    currentStrokeWidth ?? 0
+  ]);
+  map.setPaintProperty(layerId, 'circle-stroke-color', [
+    'case',
+    ['boolean', ['feature-state', 'hover'], false],
+    '#FFFFFF',
+    currentStrokeColor ?? 'transparent'
+  ]);
+
+  addHoverListeners(map, layerId, sourceLayerId);
+}
+
+/**
+ * Add a hover effect to all features in a line layer.
+ *
+ * @param map The top-level {@link maplibregl.Map} instance.
+ * @param layerId The id of the layer to instrument.
+ * @param sourceLayerId The id of the source layer to instrument. This is only
+ * necessary for {@link CartoKitLayer}s with {@link CartoKitVectorSource}s.
+ */
+export function instrumentLineHover(
+  map: maplibregl.Map,
+  layerId: string,
+  sourceLayerId?: string
+): void {
+  const currentStrokeWidth = map.getPaintProperty(layerId, 'line-width');
+  const currentStrokeColor = map.getPaintProperty(layerId, 'line-color');
+
+  map.setPaintProperty(layerId, 'line-width', [
+    'case',
+    ['boolean', ['feature-state', 'hover'], false],
+    1,
+    currentStrokeWidth ?? 0
+  ]);
+  map.setPaintProperty(layerId, 'line-color', [
+    'case',
+    ['boolean', ['feature-state', 'hover'], false],
+    '#FFFFFF',
+    currentStrokeColor ?? 'transparent'
+  ]);
+
+  addHoverListeners(map, layerId, sourceLayerId);
+}
+
+/**
+ * Add a hover effect to all features in a polygon layer.
+ *
+ * @param map The top-level {@link maplibregl.Map} instance.
+ * @param layerId The id of the layer to instrument.
+ * @param sourceLayerId The id of the source layer to instrument. This is only
+ * necessary for {@link CartoKitLayer}s with {@link CartoKitVectorSource}s.
+ */
+export function instrumentPolygonHover(
+  map: maplibregl.Map,
+  layerId: string,
+  sourceLayerId?: string
+): void {
+  map.addLayer({
+    id: `${layerId}-hover`,
+    type: 'line',
+    source: getSourceId(map, layerId),
+    'source-layer': sourceLayerId,
+    paint: {
+      'line-color': '#FFFFFF',
+      'line-width': [
+        'case',
+        ['boolean', ['feature-state', 'hover'], false],
+        1,
+        0
+      ]
+    }
+  });
+
+  addHoverListeners(map, layerId, sourceLayerId);
+}
+
+/**
+ * Wire up event listeners for hover effects.
+ *
+ * @param map The top-level {@link maplibregl.Map} instance.
+ * @param layerId The id of the layer to add event listeners to.
+ * @param sourceLayerId The id of the source layer to instrument. This is only
+ * necessary for {@link CartoKitLayer}s with {@link CartoKitVectorSource}s.
+ */
+function addHoverListeners(
+  map: maplibregl.Map,
+  layerId: string,
+  sourceLayerId?: string
+): void {
+  let hoveredFeatureId: string | null = null;
+  const canonicalLayerId = getCanonicalLayerId(layerId);
+  console.log({ layerId, canonicalLayerId });
+  const sourceId = getSourceId(map, layerId);
+
+  function onMouseMove(event: maplibregl.MapLayerMouseEvent): void {
+    if (event.features && event.features.length > 0) {
+      if (hoveredFeatureId !== null) {
+        map.setFeatureState(
+          {
+            source: sourceId,
+            id: hoveredFeatureId,
+            sourceLayer: sourceLayerId
+          },
+          { hover: false }
+        );
+      }
+
+      hoveredFeatureId = event.features[0].id?.toString() ?? null;
+
+      if (hoveredFeatureId) {
+        map.setFeatureState(
+          {
+            source: sourceId,
+            id: hoveredFeatureId,
+            sourceLayer: sourceLayerId
+          },
+          { hover: true }
+        );
+        map.getCanvas().style.cursor = 'pointer';
+      }
+
+      const currentIR = get(ir);
+
+      if (currentIR.layers[canonicalLayerId].layout.tooltip.visible) {
+        popup[canonicalLayerId] = {
+          open: true,
+          displayName: currentIR.layers[canonicalLayerId].displayName,
+          properties: event.features[0].properties
+        };
+      }
+    }
+  }
+
+  function onMouseLeave(): void {
+    if (hoveredFeatureId !== null) {
+      map.setFeatureState(
+        {
+          source: sourceId,
+          id: hoveredFeatureId,
+          sourceLayer: sourceLayerId
+        },
+        { hover: false }
+      );
+      map.getCanvas().style.cursor = '';
+    }
+
+    popup[canonicalLayerId] = {
+      open: false,
+      displayName: '',
+      properties: {}
+    };
+
+    hoveredFeatureId = null;
+  }
+
+  map.on('mousemove', layerId, onMouseMove);
+  map.on('mouseleave', layerId, onMouseLeave);
+
+  const layerListeners = listeners.value.get(layerId)!;
+
+  listeners.value.set(layerId, {
+    ...layerListeners,
+    mousemove: onMouseMove,
+    mouseleave: onMouseLeave
+  });
+}
 
 /**
  * Add a selection indicator to a feature in a point layer.
@@ -251,12 +446,12 @@ export function onFeatureLeave(
 }
 
 /**
- * Detach the select listeners registered for a set of layers.
+ * Detach the event listeners registered for a set of layers.
  *
  * @param map The top-level {@link maplibregl.Map} instance.
- * @param layerIds The ids of the layers to detach select listeners from.
+ * @param layerIds The ids of the layers to detach listeners from.
  */
-export function removeSelectListeners(
+export function removeLayerListeners(
   map: maplibregl.Map,
   layerIds: string[]
 ): void {
@@ -267,17 +462,10 @@ export function removeSelectListeners(
       return;
     }
 
-    const { click, mousemove, mouseleave } = layerListeners;
+    Object.entries(layerListeners).forEach(([event, listener]) => {
+      map.off(event as keyof LayerListeners, layerId, listener);
+    });
 
-    if (click) {
-      map.off('click', layerId, click);
-    }
-    // Leave the listeners registered by hover in place; they are detached by
-    // removeHoverListeners.
-    if (mousemove || mouseleave) {
-      listeners.value.set(layerId, { mousemove, mouseleave });
-    } else {
-      listeners.value.delete(layerId);
-    }
+    listeners.value.delete(layerId);
   });
 }
